drop extra class in basis.py
