+++ src/nemos/basis/__init__.pyi
@@ -1,12 +1,12 @@
 """Basis module stubs."""
 
 from ._basis import AdditiveBasis, MultiplicativeBasis
+from ._category import Category
 from ._custom_basis import CustomBasis
 from ._transformer_basis import TransformerBasis
 from .basis import (
     BSplineConv,
     BSplineEval,
-    Category,
     CyclicBSplineConv,
     CyclicBSplineEval,
     FourierEval,
